requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
DAWG-Python==0.7.2
docopt==0.6.2
frozenlist==1.4.0
idna==3.4
iso8601==2.0.0
multidict==6.0.4
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dotenv==1.0.0
requests==2.31.0
twitchio==2.8.1
typing_extensions==4.8.0
urllib3==2.0.5
yarl==1.9.2
